Added thank you email function
Will need to update credentials once I have a domain, but the function call just needs uncommented so when the user rsvps they will get an emailed thank you. This was mainly to test the nodemailer package, but I may use it later on for reminders as well. If not that, then reminders will be in a python script.