Server Update
Added hashing for cookies and finished the email function that sends an email to all the guests who entered an email. The reminder function takes subject and content parameters and returns the successes/total emails sent. Tested it with the admin page and everything worked properly with my main email and alias email. I also made the guests.json file more secure, requiring the request to have an admin cookie to get access.